requirements.txt
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pytesseract==0.3.7
six==1.15.0
wxPython==4.1.1
